+++ package.json
@@ -54,6 +54,7 @@
     "@coral-xyz/anchor": "^0.30.1",
     "@inquirer/prompts": "^5.5.0",
     "@mysten/sui": "^0.0.0-experimental-20250131013137",
+    "@solana/spl-token": "^0.4.12",
     "@solana/web3.js": "^1.95.4",
     "@uniswap/v2-core": "^1.0.1",
     "@uniswap/v2-periphery": "^1.1.0-beta.0",
@@ -66,6 +67,7 @@
     "elliptic": "6.5.7",
     "ethers": "^6.13.2",
     "hardhat": "^2.22.8",
+    "js-sha256": "^0.11.0",
     "wait-on": "^7.2.0"
   }
 }
