+++ packages/localnet/src/index.ts
@@ -1,17 +1,6 @@
-import * as UniswapV2Factory from "@uniswap/v2-core/build/UniswapV2Factory.json";
-import * as UniswapV2Router02 from "@uniswap/v2-periphery/build/UniswapV2Router02.json";
-import * as Custody from "@zetachain/protocol-contracts/abi/ERC20Custody.sol/ERC20Custody.json";
-import * as ERC1967Proxy from "@zetachain/protocol-contracts/abi/ERC1967Proxy.sol/ERC1967Proxy.json";
-import * as GatewayEVM from "@zetachain/protocol-contracts/abi/GatewayEVM.sol/GatewayEVM.json";
-import * as GatewayZEVM from "@zetachain/protocol-contracts/abi/GatewayZEVM.sol/GatewayZEVM.json";
-import * as SystemContract from "@zetachain/protocol-contracts/abi/SystemContractMock.sol/SystemContractMock.json";
-import * as TestERC20 from "@zetachain/protocol-contracts/abi/TestERC20.sol/TestERC20.json";
-import * as WETH9 from "@zetachain/protocol-contracts/abi/WZETA.sol/WETH9.json";
-import * as ZetaConnectorNonNative from "@zetachain/protocol-contracts/abi/ZetaConnectorNonNative.sol/ZetaConnectorNonNative.json";
-import { ethers, NonceManager, Signer } from "ethers";
+import { ethers, NonceManager } from "ethers";
 
 import { createToken } from "./createToken";
-import { deployOpts } from "./deployOpts";
 import { evmCall } from "./evmCall";
 import { evmDeposit } from "./evmDeposit";
 import { evmDepositAndCall } from "./evmDepositAndCall";
@@ -26,7 +15,8 @@ import { suiSetup } from "./suiSetup";
 import { zetachainCall } from "./zetachainCall";
 import { zetachainWithdraw } from "./zetachainWithdraw";
 import { zetachainWithdrawAndCall } from "./zetachainWithdrawAndCall";
-
+import { zetachainSetup } from "./zetachainSetup";
+import { evmSetup } from "./evmSetup";
 const FUNGIBLE_MODULE_ADDRESS = "0x735b14BB79463307AAcBED86DAf3322B1e6226aB";
 
 const foreignCoins: any[] = [];
@@ -36,226 +26,6 @@ const foreignCoins: any[] = [];
   return this.toString();
 };
 
-const prepareUniswap = async (deployer: Signer, TSS: Signer, wzeta: any) => {
-  const uniswapFactory = new ethers.ContractFactory(
-    UniswapV2Factory.abi,
-    UniswapV2Factory.bytecode,
-    deployer
-  );
-  const uniswapRouterFactory = new ethers.ContractFactory(
-    UniswapV2Router02.abi,
-    UniswapV2Router02.bytecode,
-    deployer
-  );
-
-  const uniswapFactoryInstance = await uniswapFactory.deploy(
-    await deployer.getAddress(),
-    deployOpts
-  );
-
-  const uniswapRouterInstance = await uniswapRouterFactory.deploy(
-    await uniswapFactoryInstance.getAddress(),
-    await wzeta.getAddress(),
-    deployOpts
-  );
-
-  return { uniswapFactoryInstance, uniswapRouterInstance };
-};
-
-const prepareZetaChain = async (
-  deployer: Signer,
-  wzetaAddress: any,
-  uniswapFactoryAddress: string,
-  uniswapRouterAddress: string
-) => {
-  const systemContractFactory = new ethers.ContractFactory(
-    SystemContract.abi,
-    SystemContract.bytecode,
-    deployer
-  );
-  const systemContract: any = await systemContractFactory.deploy(
-    wzetaAddress,
-    uniswapFactoryAddress,
-    uniswapRouterAddress,
-    deployOpts
-  );
-
-  const gatewayZEVMFactory = new ethers.ContractFactory(
-    GatewayZEVM.abi,
-    GatewayZEVM.bytecode,
-    deployer
-  );
-  const gatewayZEVMImpl = await gatewayZEVMFactory.deploy(deployOpts);
-
-  const gatewayZEVMInterface = new ethers.Interface(GatewayZEVM.abi);
-  const gatewayZEVMInitFragment =
-    gatewayZEVMInterface.getFunction("initialize");
-  const gatewayZEVMInitData = gatewayZEVMInterface.encodeFunctionData(
-    gatewayZEVMInitFragment as ethers.FunctionFragment,
-    [wzetaAddress, await deployer.getAddress()]
-  );
-
-  const proxyZEVMFactory = new ethers.ContractFactory(
-    ERC1967Proxy.abi,
-    ERC1967Proxy.bytecode,
-    deployer
-  );
-  const proxyZEVM = (await proxyZEVMFactory.deploy(
-    gatewayZEVMImpl.target,
-    gatewayZEVMInitData,
-    deployOpts
-  )) as any;
-
-  const gatewayZEVM = new ethers.Contract(
-    proxyZEVM.target,
-    GatewayZEVM.abi,
-    deployer
-  );
-  return { gatewayZEVM, systemContract };
-};
-
-const prepareEVM = async (deployer: Signer, tss: Signer) => {
-  const testERC20Factory = new ethers.ContractFactory(
-    TestERC20.abi,
-    TestERC20.bytecode,
-    deployer
-  );
-  const testEVMZeta = await testERC20Factory.deploy("zeta", "ZETA", deployOpts);
-
-  const gatewayEVMFactory = new ethers.ContractFactory(
-    GatewayEVM.abi,
-    GatewayEVM.bytecode,
-    deployer
-  );
-  const gatewayEVMImpl = await gatewayEVMFactory.deploy(deployOpts);
-
-  const gatewayEVMInterface = new ethers.Interface(GatewayEVM.abi);
-  const gatewayEVMInitFragment = gatewayEVMInterface.getFunction("initialize");
-  const gatewayEVMInitdata = gatewayEVMInterface.encodeFunctionData(
-    gatewayEVMInitFragment as ethers.FunctionFragment,
-    [await tss.getAddress(), testEVMZeta.target, await deployer.getAddress()]
-  );
-
-  const proxyEVMFactory = new ethers.ContractFactory(
-    ERC1967Proxy.abi,
-    ERC1967Proxy.bytecode,
-    deployer
-  );
-
-  const proxyEVM = (await proxyEVMFactory.deploy(
-    gatewayEVMImpl.target,
-    gatewayEVMInitdata,
-    deployOpts
-  )) as any;
-
-  const gatewayEVM = new ethers.Contract(
-    proxyEVM.target,
-    GatewayEVM.abi,
-    deployer
-  );
-
-  const zetaConnectorFactory = new ethers.ContractFactory(
-    ZetaConnectorNonNative.abi,
-    ZetaConnectorNonNative.bytecode,
-    deployer
-  );
-  const zetaConnectorImpl = await zetaConnectorFactory.deploy(deployOpts);
-
-  const custodyFactory = new ethers.ContractFactory(
-    Custody.abi,
-    Custody.bytecode,
-    deployer
-  );
-  const custodyImpl = await custodyFactory.deploy(deployOpts);
-
-  const zetaConnectorProxy = new ethers.Contract(
-    zetaConnectorImpl.target,
-    ZetaConnectorNonNative.abi,
-    deployer
-  );
-
-  const custodyProxy = new ethers.Contract(
-    custodyImpl.target,
-    Custody.abi,
-    deployer
-  );
-
-  // Temporarily disable
-  //
-  // await zetaConnectorProxy.initialize(
-  //   gatewayEVM.target,
-  //   testEVMZeta.target,
-  //   await tss.getAddress(),
-  //   await deployer.getAddress(),
-  //   deployOpts
-  // );
-
-  await custodyProxy.initialize(
-    gatewayEVM.target,
-    await tss.getAddress(),
-    await deployer.getAddress(),
-    deployOpts
-  );
-
-  await (gatewayEVM as any)
-    .connect(deployer)
-    .setCustody(custodyImpl.target, deployOpts);
-  await (gatewayEVM as any)
-    .connect(deployer)
-    .setConnector(zetaConnectorImpl.target, deployOpts);
-
-  return {
-    custody: custodyProxy,
-    gatewayEVM,
-    testEVMZeta,
-    zetaConnector: zetaConnectorProxy,
-  };
-};
-
-const deployProtocolContracts = async (
-  deployer: Signer,
-  tss: Signer,
-  fungibleModuleSigner: Signer
-) => {
-  const weth9Factory = new ethers.ContractFactory(
-    WETH9.abi,
-    WETH9.bytecode,
-    deployer
-  );
-  const wzeta = await weth9Factory.deploy(deployOpts);
-
-  const { uniswapFactoryInstance, uniswapRouterInstance } =
-    await prepareUniswap(deployer, tss, wzeta);
-  const { systemContract, gatewayZEVM } = await prepareZetaChain(
-    deployer,
-    wzeta.target,
-    await uniswapFactoryInstance.getAddress(),
-    await uniswapRouterInstance.getAddress()
-  );
-
-  await (wzeta as any)
-    .connect(fungibleModuleSigner)
-    .deposit({ ...deployOpts, value: ethers.parseEther("10") });
-  await (wzeta as any)
-    .connect(fungibleModuleSigner)
-    .approve(gatewayZEVM.target, ethers.parseEther("10"), deployOpts);
-  await (wzeta as any)
-    .connect(deployer)
-    .deposit({ ...deployOpts, value: ethers.parseEther("10") });
-  await (wzeta as any)
-    .connect(deployer)
-    .approve(gatewayZEVM.target, ethers.parseEther("10"), deployOpts);
-
-  return {
-    gatewayZEVM,
-    systemContract,
-    tss,
-    uniswapFactoryInstance,
-    uniswapRouterInstance,
-    wzeta,
-  };
-};
-
 export const initLocalnet = async ({
   port,
   exitOnError,
@@ -263,68 +33,71 @@ export const initLocalnet = async ({
   exitOnError: boolean;
   port: number;
 }) => {
-  let solanaAddresses: any = [];
-  if (isSolanaAvailable()) {
-    solanaAddresses = solanaSetup({
-      handlers: {
-        deposit: (args: any) =>
-          solanaDeposit({
-            args,
-            chainID: "901",
-            deployer,
-            foreignCoins,
-            fungibleModuleSigner,
-            protocolContracts,
-            provider,
-          }),
-        depositAndCall: (args: any) =>
-          solanaDepositAndCall({
-            args,
-            chainID: "901",
-            deployer,
-            foreignCoins,
-            fungibleModuleSigner,
-            protocolContracts,
-            provider,
-          }),
-      },
-    });
-  } else {
-    console.error("Solana CLI not available. Skipping setup.");
-  }
-
-  let suiEnv: any = {};
-
-  if (isSuiAvailable()) {
-    suiEnv = suiSetup({
-      handlers: {
-        deposit: (args: any) => {
-          suiDeposit({
-            args,
-            asset: ethers.ZeroAddress,
-            chainID: "103",
-            deployer,
-            foreignCoins,
-            fungibleModuleSigner,
-            protocolContracts,
-            provider,
-          });
+  const solanaSetupPromise = (async () => {
+    if (isSolanaAvailable()) {
+      return await solanaSetup({
+        handlers: {
+          deposit: (args: any) =>
+            solanaDeposit({
+              args,
+              chainID: "901",
+              deployer,
+              foreignCoins,
+              fungibleModuleSigner,
+              protocolContracts,
+              provider,
+            }),
+          depositAndCall: (args: any) =>
+            solanaDepositAndCall({
+              args,
+              chainID: "901",
+              deployer,
+              foreignCoins,
+              fungibleModuleSigner,
+              protocolContracts,
+              provider,
+            }),
         },
-        depositAndCall: (args: any) => {
-          suiDepositAndCall({
-            args,
-            asset: ethers.ZeroAddress,
-            chainID: "103",
-            deployer,
-            foreignCoins,
-            fungibleModuleSigner,
-            protocolContracts,
-            provider,
-          });
+      });
+    } else {
+      console.error("Solana CLI not available. Skipping setup.");
+    }
+  })();
+
+  const suiSetupPromise = (async () => {
+    if (isSuiAvailable()) {
+      return await suiSetup({
+        handlers: {
+          deposit: (args: any) => {
+            suiDeposit({
+              args,
+              asset: ethers.ZeroAddress,
+              chainID: "103",
+              deployer,
+              foreignCoins,
+              fungibleModuleSigner,
+              protocolContracts,
+              provider,
+            });
+          },
+          depositAndCall: (args: any) => {
+            suiDepositAndCall({
+              args,
+              asset: ethers.ZeroAddress,
+              chainID: "103",
+              deployer,
+              foreignCoins,
+              fungibleModuleSigner,
+              protocolContracts,
+              provider,
+            });
+          },
         },
-      },
-    });
-  }
+      });
+    } else {
+      console.error("Sui CLI not available. Skipping setup.");
+    }
+  })();
 
   const provider = new ethers.JsonRpcProvider(`http://127.0.0.1:${port}`);
   provider.pollingInterval = 100;
@@ -352,15 +125,19 @@ export const initLocalnet = async ({
   );
   tss = tss.connect(provider);
 
-  const protocolContracts = await deployProtocolContracts(
-    deployer,
-    tss,
-    fungibleModuleSigner
-  );
-
-  const contractsEthereum = await prepareEVM(deployer, tss);
-
-  const contractsBNB = await prepareEVM(deployer, tss);
+  const [
+    solanaEnv,
+    suiEnv,
+    contractsEthereum,
+    contractsBNB,
+    protocolContracts,
+  ] = await Promise.all([
+    solanaSetupPromise,
+    suiSetupPromise,
+    evmSetup(deployer, tss),
+    evmSetup(deployer, tss),
+    zetachainSetup(deployer, tss, fungibleModuleSigner),
+  ]);
 
   const addresses = {
     ...protocolContracts,
@@ -414,7 +191,7 @@ export const initLocalnet = async ({
       gatewayZEVM: protocolContracts.gatewayZEVM,
       provider,
       tss,
-      suiEnv: (await suiEnv).env,
+      suiEnv: suiEnv?.env,
     });
   });
 
@@ -530,9 +307,7 @@ export const initLocalnet = async ({
     }
   );
 
-  return [
-    ...(await suiEnv).addresses,
-    ...(await solanaAddresses),
+  const res = [
     ...Object.entries(protocolContracts)
       .filter(([_, value]) => value.target !== undefined)
       .map(([key, value]) => {
@@ -580,4 +355,9 @@ export const initLocalnet = async ({
       };
     }),
   ];
+
+  if (suiEnv) res.push(...suiEnv.addresses);
+  if (solanaEnv) res.push(...solanaEnv.addresses);
+
+  return res;
 };
